requirements.txt
requests
paho-mqtt>=1.6.1
Flask
python-dotenv
pytz